+++ frontend/projects/developer-registration/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { DevEditComponent } from './dev-edit/dev-edit.component';
-import { DevListComponent } from './dev-list/dev-list.component';
+import { DevEditComponent } from './devs/components/dev-edit/dev-edit.component';
+import { DevListComponent } from './devs/components/dev-list/dev-list.component';
 
 const routes: Routes = [
   { path: 'devs', component: DevListComponent },
