+++ frontend/projects/developer-registration/src/app/app.module.ts
@@ -1,32 +1,42 @@
+// Angular Dependencies
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
+import { CommonModule } from '@angular/common';
+import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { DevCardComponent } from './dev-card/dev-card.component';
+import { ReactiveFormsModule } from '@angular/forms';
+
+// Angular Material Dependencies
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-import { NavbarComponent } from './navbar/navbar.component';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatListModule } from '@angular/material/list';
-import { DevListComponent } from './dev-list/dev-list.component';
-import { DevEditComponent } from './dev-edit/dev-edit.component';
-import { ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
+
+// NgRx Dependencies
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { ReactiveComponentModule } from '@ngrx/component';
+
+// Other Dependencies
 import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
 import { faGithubSquare } from '@fortawesome/free-brands-svg-icons';
-import { HttpClientModule } from '@angular/common/http';
-import { NgxSpinnerModule } from 'ngx-spinner';
-import { CommonModule } from '@angular/common';
 import { ToastrModule } from 'ngx-toastr';
-import { StoreModule } from '@ngrx/store';
-import * as devListReducer from '../reducers/dev-list.reducer';
-import { EffectsModule } from '@ngrx/effects';
-import { DevListEffects } from '../effects/dev-list.effects';
-import { ReactiveComponentModule } from '@ngrx/component';
+import { NgxSpinnerModule } from 'ngx-spinner';
+
+// My Components
+import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { DevCardComponent } from './devs/components/dev-card/dev-card.component';
+import { NavbarComponent } from './navbar/navbar.component';
+import { DevListComponent } from './devs/components/dev-list/dev-list.component';
+import { DevEditComponent } from './devs/components/dev-edit/dev-edit.component';
+
+// My NgRx Stores
+import * as fromReducer from './devs/store/dev-list.reducer';
+import * as fromEffects from './devs/store/dev-list.effects';
 
 @NgModule({
   declarations: [
@@ -60,10 +70,10 @@ import { ReactiveComponentModule } from '@ngrx/component';
     }),
     ReactiveComponentModule,
     StoreModule.forRoot({
-      devList: devListReducer.reducer
+      devList: fromReducer.reducer
     }),
     EffectsModule.forRoot([
-      DevListEffects
+      fromEffects.DevListEffects
     ])
   ],
   providers: [],
