--- frontend/projects/developer-registration/src/app/dev-list/dev-list.component.html
@@ -1,5 +1,0 @@
-<div class="devs-container" *ngIf="(devList$ | ngrxPush) as devList">
-  <div *ngFor="let dev of devList; index as i">
-      {{ i }}. <app-dev-card [dev]="dev"></app-dev-card>
-  </div>
-</div>
