+++ frontend/projects/developer-registration/src/app/devs/components/dev-card/dev-card.component.spec.ts
