+++ frontend/projects/developer-registration/src/app/devs/components/dev-list/dev-list.component.scss
