+++ frontend/projects/developer-registration/src/app/devs/models/github-info.ts
