+++ frontend/projects/developer-registration/src/app/devs/services/github.service.ts
@@ -2,8 +2,8 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { GithubInfo } from '../entities/github-info';
-import { environment } from '../environments/environment';
+import { GithubInfo } from './../models/github-info';
+import { environment } from './../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
