+++ frontend/projects/developer-registration/src/app/devs/store/dev-list.reducer.spec.ts
