+++ frontend/projects/developer-registration/src/app/devs/store/dev-list.selectors.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import * as fromStore from './../reducers/dev-list.reducer';
+import * as fromStore from './dev-list.reducer';
 
 const devListSelector = createFeatureSelector<fromStore.DevInfoState>(fromStore.devListFeatureKey);
