--- frontend/projects/developer-registration/src/entities/dev-info.spec.ts
@@ -1,7 +1,0 @@
-import { DevInfo } from './dev-info';
-
-describe('DevInfo', () => {
-  it('should create an instance', () => {
-    expect(new DevInfo()).toBeTruthy();
-  });
-});
