--- frontend/projects/developer-registration/src/entities/github-info.spec.ts
@@ -1,7 +1,0 @@
-import { GithubInfo } from './github-info';
-
-describe('GithubInfo', () => {
-  it('should create an instance', () => {
-    expect(new GithubInfo()).toBeTruthy();
-  });
-});
